+++ src/main/java/roomescape/domain/login/controller/LoginController.java
@@ -8,9 +8,9 @@ import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RestController;
 import roomescape.domain.login.dto.LoginRequest;
-import roomescape.domain.member.Member;
-import roomescape.domain.member.MemberResponse;
-import roomescape.domain.member.MemberService;
+import roomescape.domain.member.domain.Member;
+import roomescape.domain.member.dto.MemberResponse;
+import roomescape.domain.member.service.MemberService;
 import roomescape.global.auth.CookieGenerator;
 import roomescape.global.auth.JwtTokenProvider;
