+++ src/main/java/roomescape/domain/member/dto/MemberResponse.java
@@ -1,4 +1,4 @@
-package roomescape.domain.member;
+package roomescape.domain.member.dto;
 
 public record MemberResponse(String name) {
