+++ src/main/java/roomescape/domain/member/repository/MemberRepository.java
@@ -1,7 +1,8 @@
-package roomescape.domain.member;
+package roomescape.domain.member.repository;
 
 import java.util.List;
 import java.util.Optional;
+import roomescape.domain.member.domain.Member;
 
 
 public interface MemberRepository {
