+++ src/main/java/roomescape/domain/member/repository/MemberRepositoryImpl.java
@@ -1,4 +1,4 @@
-package roomescape.domain.member;
+package roomescape.domain.member.repository;
 
 import java.util.HashMap;
 import java.util.List;
@@ -9,6 +9,8 @@ import org.springframework.jdbc.core.JdbcTemplate;
 import org.springframework.jdbc.core.RowMapper;
 import org.springframework.jdbc.core.simple.SimpleJdbcInsert;
 import org.springframework.stereotype.Repository;
+import roomescape.domain.member.domain.Member;
+import roomescape.domain.member.domain.Role;
 
 @Repository
 public class MemberRepositoryImpl implements MemberRepository {
