+++ src/main/java/roomescape/domain/reservation/controller/AdminReservationController.java
@@ -1,27 +1,23 @@
 package roomescape.domain.reservation.controller;
 
 import java.net.URI;
-import java.util.List;
 import org.springframework.http.ResponseEntity;
-import org.springframework.web.bind.annotation.DeleteMapping;
-import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RestController;
 import roomescape.domain.login.controller.AuthenticationPrincipal;
-import roomescape.domain.member.Member;
-import roomescape.domain.reservation.domain.Reservation;
+import roomescape.domain.member.domain.Member;
+import roomescape.domain.reservation.domain.reservation.Reservation;
 import roomescape.domain.reservation.dto.ReservationAddRequest;
-import roomescape.domain.reservation.service.AdminReservationService;
+import roomescape.domain.reservation.service.ReservationService;
 
 @RestController
 public class AdminReservationController {
 
-    private final AdminReservationService adminReservationService;
+    private final ReservationService reservationService;
 
-    public AdminReservationController(AdminReservationService adminReservationService) {
-        this.adminReservationService = adminReservationService;
+    public AdminReservationController(ReservationService reservationService) {
+        this.reservationService = reservationService;
     }
 
     @PostMapping("/admin/reservations")
@@ -29,7 +25,7 @@ public class AdminReservationController {
                                                       @AuthenticationPrincipal Member member) {
         reservationAddRequest = new ReservationAddRequest(reservationAddRequest.date(), reservationAddRequest.timeId(),
                 reservationAddRequest.themeId(), member.getId());
-        Reservation reservation = adminReservationService.addReservation(reservationAddRequest);
+        Reservation reservation = reservationService.addReservation(reservationAddRequest);
         return ResponseEntity.created(URI.create("/reservation/" + reservation.getId())).body(reservation);
     }
 }
