+++ src/main/java/roomescape/domain/reservation/dto/ReservationTimeAddRequest.java
@@ -1,7 +1,7 @@
-package roomescape.domain.reservationTime.dto;
+package roomescape.domain.reservation.dto;
 
 import java.time.LocalTime;
-import roomescape.domain.reservationTime.domain.ReservationTime;
+import roomescape.domain.reservation.domain.reservationTim.ReservationTime;
 
 public record ReservationTimeAddRequest(LocalTime startAt) {
     public ReservationTime toEntity() {
