+++ src/main/java/roomescape/domain/reservation/service/AdminReservationTimeService.java
@@ -1,11 +1,11 @@
-package roomescape.domain.reservationTime.service;
+package roomescape.domain.reservation.service;
 
 import java.util.List;
 import org.springframework.stereotype.Service;
+import roomescape.domain.reservation.domain.reservationTim.ReservationTime;
+import roomescape.domain.reservation.dto.ReservationTimeAddRequest;
+import roomescape.domain.reservation.repository.ReservationTimeRepository;
 import roomescape.global.exception.ClientIllegalArgumentException;
-import roomescape.domain.reservationTime.domain.ReservationTime;
-import roomescape.domain.reservationTime.dto.ReservationTimeAddRequest;
-import roomescape.domain.reservationTime.repository.ReservationTimeRepository;
 
 @Service
 public class AdminReservationTimeService {
