--- src/main/java/roomescape/domain/theme/controller/AdminThemeController.java
@@ -1,40 +1,0 @@
-package roomescape.domain.theme.controller;
-
-import java.net.URI;
-import java.util.List;
-import org.springframework.http.ResponseEntity;
-import org.springframework.web.bind.annotation.DeleteMapping;
-import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.PathVariable;
-import org.springframework.web.bind.annotation.PostMapping;
-import org.springframework.web.bind.annotation.RequestBody;
-import org.springframework.web.bind.annotation.RestController;
-import roomescape.domain.theme.domain.Theme;
-import roomescape.domain.theme.dto.ThemeAddRequest;
-import roomescape.domain.theme.service.AdminThemeService;
-
-@RestController
-public class AdminThemeController {
-    private final AdminThemeService adminThemeService;
-
-    public AdminThemeController(AdminThemeService adminThemeService) {
-        this.adminThemeService = adminThemeService;
-    }
-
-    @GetMapping("/themes")
-    public ResponseEntity<List<Theme>> getThemeList() {
-        return ResponseEntity.ok(adminThemeService.findAllTheme());
-    }
-
-    @PostMapping("/themes")
-    public ResponseEntity<Theme> addTheme(@RequestBody ThemeAddRequest themeAddRequest) {
-        Theme theme = adminThemeService.addTheme(themeAddRequest);
-        return ResponseEntity.created(URI.create("/themes" + theme.getId())).body(theme);
-    }
-
-    @DeleteMapping("/themes/{id}")
-    public ResponseEntity<Void> deleteTheme(@PathVariable("id") Long id) {
-        adminThemeService.removeTheme(id);
-        return ResponseEntity.noContent().build();
-    }
-}
