+++ src/main/java/roomescape/domain/theme/controller/ThemeController.java
@@ -1,10 +1,16 @@
 package roomescape.domain.theme.controller;
 
+import java.net.URI;
 import java.util.List;
 import org.springframework.http.ResponseEntity;
+import org.springframework.web.bind.annotation.DeleteMapping;
 import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.PathVariable;
+import org.springframework.web.bind.annotation.PostMapping;
+import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RestController;
 import roomescape.domain.theme.domain.Theme;
+import roomescape.domain.theme.dto.ThemeAddRequest;
 import roomescape.domain.theme.service.ThemeService;
 
 @RestController
@@ -14,6 +20,23 @@ public class ThemeController {
 
     public ThemeController(ThemeService themeService) {
         this.themeService = themeService;
+    }
+
+    @GetMapping("/themes")
+    public ResponseEntity<List<Theme>> getThemeList() {
+        return ResponseEntity.ok(themeService.findAllTheme());
+    }
+
+    @PostMapping("/themes")
+    public ResponseEntity<Theme> addTheme(@RequestBody ThemeAddRequest themeAddRequest) {
+        Theme theme = themeService.addTheme(themeAddRequest);
+        return ResponseEntity.created(URI.create("/themes" + theme.getId())).body(theme);
+    }
+
+    @DeleteMapping("/themes/{id}")
+    public ResponseEntity<Void> deleteTheme(@PathVariable("id") Long id) {
+        themeService.removeTheme(id);
+        return ResponseEntity.noContent().build();
     }
 
     @GetMapping("/theme-ranking")
