+++ src/test/java/roomescape/domain/login/repository/MemberRepositoryImplTest.java
@@ -1,7 +1,7 @@
 package roomescape.domain.login.repository;
 
 import static org.assertj.core.api.Assertions.assertThat;
-import static roomescape.domain.member.Role.MEMBER;
+import static roomescape.domain.member.domain.Role.MEMBER;
 import static roomescape.fixture.MemberFixture.ADMIN_MEMBER;
 
 import org.junit.jupiter.api.AfterEach;
@@ -11,9 +11,9 @@ import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.jdbc.core.JdbcTemplate;
 import roomescape.RepositoryTest;
-import roomescape.domain.member.Member;
-import roomescape.domain.member.MemberRepository;
-import roomescape.domain.member.MemberRepositoryImpl;
+import roomescape.domain.member.domain.Member;
+import roomescape.domain.member.repository.MemberRepository;
+import roomescape.domain.member.repository.MemberRepositoryImpl;
 
 class MemberRepositoryImplTest extends RepositoryTest {
