+++ src/test/java/roomescape/domain/member/RoleTest.java
@@ -4,6 +4,7 @@ import static org.assertj.core.api.Assertions.assertThat;
 
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
+import roomescape.domain.member.domain.Role;
 
 class RoleTest {
