+++ src/test/java/roomescape/domain/reservation/domain/ReservationDateTest.java
@@ -6,6 +6,7 @@ import static roomescape.fixture.LocalDateFixture.AFTER_ONE_DAYS_DATE;
 
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
+import roomescape.domain.reservation.domain.reservation.ReservationDate;
 import roomescape.global.exception.ClientIllegalArgumentException;
 
 class ReservationDateTest {
