+++ src/test/java/roomescape/domain/reservationTime/service/AdminReservationTimeServiceTest.java
@@ -8,8 +8,9 @@ import java.time.LocalTime;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
-import roomescape.domain.reservationTime.domain.ReservationTime;
-import roomescape.domain.reservationTime.dto.ReservationTimeAddRequest;
+import roomescape.domain.reservation.domain.reservationTim.ReservationTime;
+import roomescape.domain.reservation.dto.ReservationTimeAddRequest;
+import roomescape.domain.reservation.service.AdminReservationTimeService;
 import roomescape.global.exception.ClientIllegalArgumentException;
 
 class AdminReservationTimeServiceTest {
