+++ src/test/java/roomescape/domain/reservationTime/service/FakeReservationTimeRepository.java
@@ -6,8 +6,8 @@ import java.util.List;
 import java.util.Map;
 import java.util.Optional;
 import java.util.concurrent.atomic.AtomicLong;
-import roomescape.domain.reservationTime.domain.ReservationTime;
-import roomescape.domain.reservationTime.repository.ReservationTimeRepository;
+import roomescape.domain.reservation.domain.reservationTim.ReservationTime;
+import roomescape.domain.reservation.repository.ReservationTimeRepository;
 
 public class FakeReservationTimeRepository implements ReservationTimeRepository {
